Fix 32 compile errors: Ghidra decompilation syntax issues
Fixed common patterns from Ghidra output:
- Operator precedence: ptr + expr << n -> ptr + (expr << n)
- Invalid dereference: *((int)ptr << n) -> (*ptr << n)
- Missing function parameters: added param_5, param_6, param_7
- Function pointer casts: (*array[idx])() -> (*(int(*)())array[idx])()

Compile errors reduced from 67 to 35.

+++ src/FUN_06016DD8.c
@@ -120,7 +120,7 @@ void FUN_06016dd8(param_1)
 
     *(short **)(&stack0xfffffff4 + iVar3) = psVar8;
 
-    iVar5 = (*(int(*)())puVar1)((int)*(int)psVar8 << 0x10,*(int *)((int)&local_20 + iVar3));
+    iVar5 = (*(int(*)())puVar1)((int)*psVar8 << 0x10,*(int *)((int)&local_20 + iVar3));
 
     **(int **)((int)aiStack_28 + iVar3 + 4) = iVar5 + *(int *)((int)aiStack_28 + iVar3);
