+++ src/FUN_060196A4.c
@@ -27,15 +27,15 @@ void FUN_060196a4()
 
     (*(int(*)())0x06028400)(0xc,*(int *)
 
-                    (0x06063750 + (unsigned int)*(unsigned short *)(0x06049AF4 + (char)*((int)(int)puVar3 << 1)) << 3)
+                    (0x06063750 + (unsigned int)*(unsigned short *)(0x06049AF4 + (char)(*puVar3 << 1)) << 3)
 
-               ,((unsigned int)(unsigned char)((char *)(0x06049AEC + (char)*((int)(int)puVar3 << 1)))[1] * 0x40 +
+               ,((unsigned int)(unsigned char)((char *)(0x06049AEC + (char)(*puVar3 << 1)))[1] * 0x40 +
 
-                (unsigned int)(unsigned char)((int *)0x06049AEC)[(char)*((int)(int)puVar3 << 1)]) << 1,
+                (unsigned int)(unsigned char)((int *)0x06049AEC)[(char)(*puVar3 << 1)]) << 1,
 
                *(int *)((int)(0x06063750 +
 
-                             (unsigned int)*(unsigned short *)(0x06049AF4 + (char)*((int)(int)puVar3 << 1)) << 3) + 4) +
+                             (unsigned int)*(unsigned short *)(0x06049AF4 + (char)(*puVar3 << 1)) << 3) + 4) +
 
                (int)(short)PTR_DAT_06019764);
 
@@ -99,7 +99,7 @@ void FUN_060196a4()
 
   *puVar2 = *puVar2 + '\x01';
 
-  (*(int(*)())(*(int *)((char)*((int)(int)puVar3 << 2) + 0x0605D250)))();
+  (*(int(*)())(*(int *)((char)(*puVar3 << 2) + 0x0605D250)))();
 
   return;
